BSL: Bayesian Synthetic Likelihood with Graphical Lasso
Bayesian synthetic likelihood (BSL, Price et al. (2018) <doi:10.1080/10618600.2017.1302882>) is an alternative to standard, non-parametric approximate Bayesian computation (ABC). BSL assumes a multivariate normal distribution for the summary statistic likelihood and it is suitable when the distribution of the model summary statistics is sufficiently regular. This package provides a Metropolis Hastings Markov chain Monte Carlo implementation of BSL and BSL with graphical lasso (BSLasso, An et al. (2018) <https://eprints.qut.edu.au/102263/>), which is computationally more efficient when the dimension of the summary statistic is high. Extensions to this package are planned.